+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/settings/production.py
@@ -1,4 +1,4 @@
-from .base import *
+from .base import *  # noqa
 
 
 # Disable debug mode
@@ -44,16 +44,16 @@ COMPRESS_OFFLINE = True
 # Setup for CAS
 ENABLE_SSO = True
 
-MIDDLEWARE_CLASSES += (
+MIDDLEWARE_CLASSES += [
     'molo.core.middleware.MoloCASMiddleware',
     'molo.core.middleware.Custom403Middleware',
-)
+]
 
 
-AUTHENTICATION_BACKENDS = (
+AUTHENTICATION_BACKENDS = [
     'django.contrib.auth.backends.ModelBackend',
     'molo.core.backends.MoloCASBackend',
-)
+]
 
 CAS_SERVER_URL = ''
 CAS_ADMIN_PREFIX = '/admin/'
@@ -61,6 +61,6 @@ LOGIN_URL = '/accounts/login/'
 CAS_VERSION = '3'
 
 try:
-    from .local import *
+    from .local import *  # noqa
 except ImportError:
     pass